Intro to Machine Learning With Pytorch/Image Classifier/requirements.txt
argparse
sys
numpy
torch
torchvision
seaborn
matplotlib
Pillow